Release v1.0.0
This action will automatically bring a GitHub Action’s latest tags
to the major and minor of the most recent tag.

Example: after pushing a v7.4.1 tag for a GitHub Action, this will
automatically sync the v7 and v7.4 tags of that repository, saving
you the trouble of doing it manually.